integer type fix

+++ 0x0B-malloc_free/1-strdup.c
@@ -8,7 +8,7 @@
 
 char *_strdup(char *str)
 {
-	unsigned long int i;
+	int i;
 	int size;
 	char *temp;
